backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
numpy>=2.0.0
nilearn>=0.10.0
nibabel>=5.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
scipy>=1.13.0

# Vision LLM clients (used by services/vision_client.py + services/comparison.py).
# httpx above already covers the raw HTTP path for Anthropic/OpenAI/Gemini, but
# pulling these SDKs in keeps the door open for native streaming / file APIs.
anthropic>=0.40.0
google-generativeai>=0.8.0

# Optional: used by services/vision_client.py for in-process frame extraction.
# Falls back to ffmpeg subprocess if absent, so this is not strictly required.
opencv-python-headless>=4.9.0

# Embedding proxy (services/embedding_proxy) + training-pair YAML loader.
sentence-transformers>=2.7.0
PyYAML>=6.0
